Using mumps on ci.
First attempt to compile dynamical planner on actions.

Using another strategy to checkout.

Fixed bug in CI and added missing dependency.

Added some missing dependency.

Attempt to fix missing ffmpeg.

Added more ffmpeg deps.

Shipping shlibpp if not found in levi.

Attempt to run SolverForComparisonUnitTest with xvfb-run

Attempt to download artifacts

Moved download of artifacts to the end

We need to "Upload" artifacts, not downloading them.

Tryig to install also libopenblas-dev

Using a different checkout strategy for Ipopt and its deps.

Attempt to change Actions governor.

Revert "Attempt to change Actions governor."

This reverts commit 13aad8f8685ab8d9d2014e9129a842c97d374e3d.

Not possible with GHA, worth a shot though.

Added the possibility to set some of the SolverForComparisonsTest parameters from cmd

Attempt to create an artifact with the build folder.

Attempt to fix CI failure on different build folder by removing the coinbrew install

It should not be needed anyway

Fixing CI compilation and attempt to run tests in parallel

Fixed a path when unzipping the build folder.

Fixing the generation of the build zip.

Adding system dependencies when running tests.

Fetching the repository in the test

Fix the path to the directory to zip

Fixing a couple of bugs in the tests when using mumps with the hyperbolic method.

Fixing path for saving artifacts.

Reducing mumps pivot tolerance.

Attempt to run parameter variation test in Actions

Using arrays with include does not work